<FIX> Fix some issues when joining
1. Add a random backoff timer when sending joinAck back to the new node. Previously
I noticed some packet losses when two nodes replying back to the new node with their
joinAck messages. This might be due to the packet collison when two nodes receive
the join request and reply back almost simultaneously. In my test, it is usually the
node with better RSSI value win the collision (The other node's message gets lost).
Thus, a backoff time should solve this collision issue.

2. Fix the RSSI value. It should always be negative as it is in nature.

+++ MessageProcessor.h
@@ -35,7 +35,7 @@ public:
     /**
      * For every message receveid, there will be an RSSI value associated
      */
-    uint8_t rssi;
+    int rssi;
 
     GenericMessage(byte type, byte* srcAddr, byte* destAddr);
     // return number of bytes sent
